feat(watcher): tend the prediction desks each tick
- Refreshes every busy desk's remembered price. A window's makers stop
  quoting long before it expires, and a desk with no price to report makes
  Arena value a real holding at nothing for the tail of a fight.
- Frees a desk only once its own window has expired and nothing is
  running, collecting winnings first. Releasing merely because no fight is
  live would free a desk claimed moments before its fight begins, letting
  something re-point it underneath.
- Does nothing when no desks are in the manifest, so existing deployments
  behave exactly as before.

+++ coliseum/scripts/deploy-event-desks.ts
@@ -76,7 +76,7 @@ async function main() {
   console.log(`Refilling treasury with ${formatUnits(total, 6)} from the faucet...`);
   const refill = await treasury.write.refill([total]);
   await pub.waitForTransactionReceipt({ hash: refill });
-  console.log(`Treasury balance: ${formatUnits(await treasury.read.balance(), 6)}`);
+  console.log(`Treasury balance: ${formatUnits((await treasury.read.balance()) as bigint, 6)}`);
 
   // Desks are NOT funded here. `fund` puts collateral into the bound market's
   // vault, and an unbound desk has no vault to put it in — so funding happens
